+++ app/models/categories.rb
