+++ app/models/departments.rb
