+++ app/models/products.rb
