+++ app/views/categories.erb
