+++ config.ru
@@ -1,3 +1,7 @@
-require './app'
+require './app/controllers/application_controller'
+require './app/controllers/departments_controller'
+require './app/controllers/categories_controller'
 
-run App
+use DepartmentsController
+use CategoriesController
+run ApplicationController
